requirements.txt
streamlit==1.32.0
reportlab==4.1.0
Pillow==10.2.0
openai==1.12.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
matplotlib==3.8.3
requests==2.31.0 